Refactor hooks, interfaces, and  components.
